changed int to long for the BitcoinWalletInformation

+++ src/Carbonara/Models/BitcoinWalletInformation.cs
@@ -6,28 +6,28 @@ namespace Carbonara.Models
     {
         public string hash160 { get; set; }
         public string address { get; set; }
-        public int n_tx { get; set; }
-        public int total_received { get; set; }
-        public int total_sent { get; set; }
-        public int final_balance { get; set; }
+        public long n_tx { get; set; }
+        public long total_received { get; set; }
+        public long total_sent { get; set; }
+        public long final_balance { get; set; }
         public List<Tx> txs { get; set; }
     }
 
     public class SpendingOutpoint
 {
-    public int tx_index { get; set; }
-    public int n { get; set; }
+    public long tx_index { get; set; }
+    public long n { get; set; }
 }
 
 public class PrevOut
 {
     public bool spent { get; set; }
     public List<SpendingOutpoint> spending_outpoints { get; set; }
-    public int tx_index { get; set; }
-    public int type { get; set; }
+    public long tx_index { get; set; }
+    public long type { get; set; }
     public string addr { get; set; }
     public long value { get; set; }
-    public int n { get; set; }
+    public long n { get; set; }
     public string script { get; set; }
 }
 
@@ -41,38 +41,38 @@ public class Input
 
 public class SpendingOutpoint2
 {
-    public int tx_index { get; set; }
-    public int n { get; set; }
+    public long tx_index { get; set; }
+    public long n { get; set; }
 }
 
 public class Out
 {
     public bool spent { get; set; }
     public List<SpendingOutpoint2> spending_outpoints { get; set; }
-    public int tx_index { get; set; }
-    public int type { get; set; }
+    public long tx_index { get; set; }
+    public long type { get; set; }
     public string addr { get; set; }
     public long value { get; set; }
-    public int n { get; set; }
+    public long n { get; set; }
     public string script { get; set; }
 }
 
 public class Tx
 {
-    public int ver { get; set; }
+    public long ver { get; set; }
     public List<Input> inputs { get; set; }
-    public int weight { get; set; }
-    public int block_height { get; set; }
+    public long weight { get; set; }
+    public long block_height { get; set; }
     public string relayed_by { get; set; }
     public List<Out> @out { get; set; }
-    public int lock_time { get; set; }
-    public int result { get; set; }
-    public int size { get; set; }
-    public int block_index { get; set; }
-    public int time { get; set; }
-    public int tx_index { get; set; }
-    public int vin_sz { get; set; }
+    public long lock_time { get; set; }
+    public long result { get; set; }
+    public long size { get; set; }
+    public long block_index { get; set; }
+    public long time { get; set; }
+    public long tx_index { get; set; }
+    public long vin_sz { get; set; }
     public string hash { get; set; }
-    public int vout_sz { get; set; }
+    public long vout_sz { get; set; }
 }
 }
